Update ruleset param validator to consider OperationContextParams
